refactor(order): OrderPostItemReq price 타입 Long에서 Integer 변경

+++ order-service/src/main/java/com/green/eats/order/application/model/OrderPostItemReq.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class OrderPostItemReq {
     private Long menuId; // 주문할 메뉴 ID
     private Integer quantity; // 수량
-    private Long price; // 단가
+    private Integer price; // 단가
 }
